refactor: extract fetchArtistFans into lib/fans
Move the inline fetchFans function out of hooks/useArtistFans.ts and into
lib/fans/fetchArtistFans.ts, matching the pattern established by
lib/artists/fetchArtists.ts. The hook keeps the Privy access-token plumbing
and passes the token into the lib function.

Types (Social, FansResponse, FansError) now live alongside the fetcher and
are re-exported from the hook for backward compatibility with existing
consumers (components/Fans/*).
